backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["Firmino Neto <firminoneto11@hotmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
orm = "0.3.1"
fastapi = "0.85.0"
python-decouple = "3.6"
httpx = "0.23.0"
uvicorn = "0.18.3"
pandas = "1.5.0"
asyncpg = "0.26.0"
psycopg2-binary = "2.9.4"

[tool.poetry.dev-dependencies]
black = "22.8.0"
ipython = "8.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
